Move verification methods and members to TestWithMiniDFSCluster
